Refactor [#154] 홈 화면 통합 VC 구현
